chronosServer/requirements.txt
fastapi==0.118.0
uvicorn==0.32.1
supabase==2.20.0
python-dotenv==1.0.1
pydantic==2.10.5
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.2.1
httpx==0.27.2
python-multipart==0.0.20
pytz==2024.2
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.184.0